requirements.txt
pandas
pyTelegramBotAPI
SQLAlchemy
psycopg2-binary
selenium
beautifulsoup4
PyYAML
Pillow
timm
sklearn
